% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilFoodWebs_SOFIA-package.R
\docType{package}
\name{SoilFoodWebs_SOFIA-package}
\alias{SoilFoodWebs_SOFIA}
\alias{SoilFoodWebs_SOFIA-package}
\title{SoilFoodWebs_SOFIA: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MikeHedde/SoilFoodWebs_SOFIA}
  \item Report bugs at \url{https://github.com/MikeHedde/SoilFoodWebs_SOFIA/issues}
}

}
\author{
\strong{Maintainer}: Mike Hedde \email{mickael.hedde@inrae.fr} (\href{https://orcid.org/0000-0002-6733-3622}{ORCID}) [copyright holder]

}
\keyword{internal}
